# Optimize matching performance on average and for specific edge-cases