--- a/internal/own/codeowners/find_owners_test.go
+++ b/internal/own/codeowners/find_owners_test.go
@@ -1,6 +1,7 @@
 package codeowners_test
 
 import (
+	"fmt"
 	"reflect"
 	"testing"
 
@@ -360,6 +361,79 @@ func BenchmarkOwnersMismatchAbsoluteGlob(b *testing.B) {
 			{Pattern: pattern, Owner: owner},
 		},
 	})
+	// Warm cache.
+	for _, path := range paths {
+		rs.FindOwners(path)
+	}
+
+	for i := 0; i < b.N; i++ {
+		rs.FindOwners(pattern)
+	}
+}
+
+func BenchmarkOwnersMatchMultiHole(b *testing.B) {
+	pattern := "/main/**/foo/**/*.md"
+	paths := []string{
+		"/main/src/foo/bar/README.md",
+	}
+	owner := []*codeownerspb.Owner{
+		{Handle: "foo"},
+	}
+	rs := codeowners.NewRuleset(&codeownerspb.File{
+		Rule: []*codeownerspb.Rule{
+			{Pattern: pattern, Owner: owner},
+		},
+	})
+	// Warm cache.
+	for _, path := range paths {
+		rs.FindOwners(path)
+	}
+
+	for i := 0; i < b.N; i++ {
+		rs.FindOwners(pattern)
+	}
+}
+
+func BenchmarkOwnersMismatchMultiHole(b *testing.B) {
+	pattern := "/main/**/foo/**/*.md"
+	paths := []string{
+		"/main/src/foo/bar/README.txt",
+	}
+	owner := []*codeownerspb.Owner{
+		{Handle: "foo"},
+	}
+	rs := codeowners.NewRuleset(&codeownerspb.File{
+		Rule: []*codeownerspb.Rule{
+			{Pattern: pattern, Owner: owner},
+		},
+	})
+	// Warm cache.
+	for _, path := range paths {
+		rs.FindOwners(path)
+	}
+
+	for i := 0; i < b.N; i++ {
+		rs.FindOwners(pattern)
+	}
+}
+
+func BenchmarkOwnersMatchLiteralLargeRuleset(b *testing.B) {
+	pattern := "/main/src/foo/bar/README.md"
+	paths := []string{
+		"/main/src/foo/bar/README.md",
+	}
+	owner := []*codeownerspb.Owner{
+		{Handle: "foo"},
+	}
+	f := &codeownerspb.File{
+		Rule: []*codeownerspb.Rule{
+			{Pattern: pattern, Owner: owner},
+		},
+	}
+	for i := 0; i < 10000; i++ {
+		f.Rule = append(f.Rule, &codeownerspb.Rule{Pattern: fmt.Sprintf("%s-%d", pattern, i), Owner: owner})
+	}
+	rs := codeowners.NewRuleset(f)
 	// Warm cache.
 	for _, path := range paths {
 		rs.FindOwners(path)