Benchmark tests included!

**Before**
```
BenchmarkOwnersMatchLiteral-10                   5075626               235.8 ns/op
BenchmarkOwnersMatchRelativeGlob-10              6787929               171.9 ns/op
BenchmarkOwnersMatchAbsoluteGlob-10              6564476               182.4 ns/op
BenchmarkOwnersMismatchLiteral-10                5322088               224.8 ns/op
BenchmarkOwnersMismatchRelativeGlob-10           7058532               170.5 ns/op
BenchmarkOwnersMismatchAbsoluteGlob-10           6372730               185.2 ns/op
```

**After**
```
BenchmarkOwnersMatchLiteral-10                  128574664                9.272 ns/op
BenchmarkOwnersMatchRelativeGlob-10             23077292                51.79 ns/op
BenchmarkOwnersMatchAbsoluteGlob-10             18436671                61.25 ns/op
BenchmarkOwnersMismatchLiteral-10               131070199                9.101 ns/op
BenchmarkOwnersMismatchRelativeGlob-10          23067070                52.58 ns/op
BenchmarkOwnersMismatchAbsoluteGlob-10          19928616                61.10 ns/op
```

## Test plan

Test suite still passes and added benchmark tests to verify performance improvements.